+++ jms/components/org.wso2.carbon.transport.jms/src/main/java/org/wso2/carbon/transport/jms/utils/JMSUtils.java
@@ -22,6 +22,8 @@ import org.apache.commons.logging.LogFactory;
 import org.wso2.carbon.messaging.CarbonMessage;
 import org.wso2.carbon.messaging.TextJMSCarbonMessage;
 
+import java.util.Iterator;
+import java.util.Map;
 import java.util.Properties;
 
 import javax.jms.Destination;
@@ -149,5 +151,76 @@ public class JMSUtils {
             log.error("Error while chaging the jms message to carbon message");
         }
         return jmsCarbonMessage;
+    }
+
+    public static void setTransportHeaders(Message message, Map<String, Object> headerMap) {
+        try {
+            if (headerMap != null) {
+                Iterator iterator = headerMap.keySet().iterator();
+
+                while (true) {
+                    String name;
+                    do {
+                        if (!iterator.hasNext()) {
+                            return;
+                        }
+
+                        Object headerName = iterator.next();
+                        name = (String) headerName;
+                    } while (name.startsWith("JMSX") && !name.equals("JMSXGroupID") &&
+                             !name.equals("JMSXGroupSeq"));
+
+                    if ("JMS_COORELATION_ID".equals(name)) {
+                        message.setJMSCorrelationID((String) headerMap.get("JMS_COORELATION_ID"));
+                    } else {
+                        Object value;
+                        if ("JMS_DELIVERY_MODE".equals(name)) {
+                            value = headerMap.get("JMS_DELIVERY_MODE");
+                            if (value instanceof Integer) {
+                                message.setJMSDeliveryMode(((Integer) value).intValue());
+                            } else if (value instanceof String) {
+                                try {
+                                    message.setJMSDeliveryMode(Integer.parseInt((String) value));
+                                } catch (NumberFormatException var8) {
+
+                                }
+                            } else {
+
+                            }
+                        } else if ("JMS_EXPIRATION".equals(name)) {
+                            message.setJMSExpiration(
+                                    Long.parseLong((String) headerMap.get("JMS_EXPIRATION")));
+                        } else if ("JMS_MESSAGE_ID".equals(name)) {
+                            message.setJMSMessageID((String) headerMap.get("JMS_MESSAGE_ID"));
+                        } else if ("JMS_PRIORITY".equals(name)) {
+                            message.setJMSPriority(
+                                    Integer.parseInt((String) headerMap.get("JMS_PRIORITY")));
+                        } else if ("JMS_TIMESTAMP".equals(name)) {
+                            message.setJMSTimestamp(
+                                    Long.parseLong((String) headerMap.get("JMS_TIMESTAMP")));
+                        } else if ("JMS_MESSAGE_TYPE".equals(name)) {
+                            message.setJMSType((String) headerMap.get("JMS_MESSAGE_TYPE"));
+                        } else {
+                            value = headerMap.get(name);
+                            if (value instanceof String) {
+                                message.setStringProperty(name, (String) value);
+                            } else if (value instanceof Boolean) {
+                                message.setBooleanProperty(name, ((Boolean) value).booleanValue());
+                            } else if (value instanceof Integer) {
+                                message.setIntProperty(name, ((Integer) value).intValue());
+                            } else if (value instanceof Long) {
+                                message.setLongProperty(name, ((Long) value).longValue());
+                            } else if (value instanceof Double) {
+                                message.setDoubleProperty(name, ((Double) value).doubleValue());
+                            } else if (value instanceof Float) {
+                                message.setFloatProperty(name, ((Float) value).floatValue());
+                            }
+                        }
+                    }
+                }
+            }
+        } catch (JMSException exception) {
+
+        }
     }
 }
